enviroment core
